StringBuilder
Hi , we will see about the StringBuilder , in Java String is immutable (Unchangeable) but StringBuilder is mutable (changeable) .
if we want to do some modification in our string then we used the StringBuilder. 
In this program we will see how we can use StringBuilder and some inbuild functions of StringBuilder.