Stack implementation using Linked list
In this program we will see the concept about stack using Linked list.

Stack : Stack is a linear data structure that is used to store the collection of objects. It's based on the last in first out (LIFO) concept.
In stack we have 3 important methods :
1. Push : When we need to add data in stack then we use push operation.
2. Pop : When we need to remove data from stack then we use pop operation.
3. Peek : When we need to check or see our top data in stack then we use peek operator.

NOTE : Head is our peak in linked list.