Stack Collection framework
You can see in the program i used stack collection it's just like a ArrayList. we have by default method in stack collection framwork like push() , pop(), peek() . so we don't need to create them manually.
OR 
In this program i push some values in stack then i create 2 methods :

pushatbottom() : In this function i added the value in first position in stack not in peek.
example - we have stack with data 1,2,3 so when we push the data in stack so it will add after 4 but in this method it will push before 1.

reversestack() : In this method i reversed my stack and store in temporary stack.