Queue implementation by linked list
In this program i implemented queue using Linked list. and i created 3 methods add, remove and peek or print method.
In Queue we followed FIFO first in first out. we have front and rear.
if we need to add data in queue then we need to shift our tail to next.
and if we need to remove data from our queue then we need to shift head to the next and java garbage collector removed the head data of queue.